PlayerInfo response
